Bridge Module to CUDARuntime with context control

+++ Sources/CUDARuntime/Device.swift
@@ -12,7 +12,7 @@ public struct Device : Equatable {
 
     public typealias Properties = cudaDeviceProp
 
-    public let index: Int32
+    public let index: Int
 
     public static var count: Int {
         var deviceCount: Int32 = 0
@@ -27,19 +27,20 @@ public struct Device : Equatable {
             return Device(assumingIndex: index)
         }
         set {
-            !!cudaSetDevice(newValue.index)
+            !!cudaSetDevice(Int32(newValue.index))
         }
     }
 
+    /// Unsafely assume the index handle to be the device index
     internal init(assumingIndex index: Int32) {
-        self.index = index
+        self.index = Int(index)
     }
 
     public init?(atIndex index: Int) {
         guard index >= 0 && index < Device.count else {
             return nil
         }
-        self.index = Int32(index)
+        self.index = index
     }
 
     public init?(withProperties properties: Properties) {
@@ -47,7 +48,7 @@ public struct Device : Equatable {
         var properties = properties
         do {
             try ensureSuccess(cudaChooseDevice(&index, &properties))
-            self.index = index
+            self.index = Int(index)
         } catch {
             return nil
         }
@@ -55,15 +56,15 @@ public struct Device : Equatable {
 
     public var properties: Properties {
         var prop = Properties()
-        !!cudaGetDeviceProperties(&prop, index)
+        !!cudaGetDeviceProperties(&prop, Int32(index))
         return prop
     }
 
     public var computeCapability: (major: Int, minor: Int) {
         var major: Int32 = 0
-        !!cudaDeviceGetAttribute(&major, cudaDevAttrComputeCapabilityMajor, index)
+        !!cudaDeviceGetAttribute(&major, cudaDevAttrComputeCapabilityMajor, Int32(index))
         var minor: Int32 = 0
-        !!cudaDeviceGetAttribute(&minor, cudaDevAttrComputeCapabilityMinor, index)
+        !!cudaDeviceGetAttribute(&minor, cudaDevAttrComputeCapabilityMinor, Int32(index))
         return (major: Int(major), minor: Int(minor))
     }
